+++ service/parse.py
@@ -1,16 +1,15 @@
-import model_conesearch
-import model_cross_match
+from service.model_conesearch import ModelConesearch
+from service.model_cross_match import ModelCrossMatch
 def parse_conesearch(match, catalog_columns, column_units):
 
-    model = model_conesearch(match, catalog_columns,column_units)
+    model = ModelConesearch(match, catalog_columns,column_units)
 
     return model.return_format()
 
 
-    return
 
 
 def parse_crossmatch(match, catalog, ra, dec, catalog_columns, column_units,map_ra_dec):
-    model = model_cross_match(match, catalog, ra, dec, catalog_columns, column_units, map_ra_dec)
+    model = ModelCrossMatch(match, catalog, ra, dec, catalog_columns, column_units, map_ra_dec)
 
     return model.return_format()
